hopefully the last single-to-double conversions were hiding out in libdycore
git-svn-id: http://oceans11.lanl.gov/svn/PISCEES/trunk@645 42a2fd4b-9c52-44de-a8a9-1772c67210cd

Former-commit-id: 8f7213e16983ccef55e086a78c2492cfbce1bb74

+++ libdycore/BISICLES/BisiclesToGlimmer.cpp
@@ -33,7 +33,7 @@ BisiclesToGlimmer::initDyCore(const char * input_fname)
 
 // updates cur_time_yr to match time update in dycore
 int
-BisiclesToGlimmer::runDyCore(float& cur_time_yr, const float time_inc_yr)
+BisiclesToGlimmer::runDyCore(double& cur_time_yr, const double time_inc_yr)
 {
   cout << "In BISICLES runDyCore" << endl;
   bike_driver_run(this,cur_time_yr,time_inc_yr);
